pyproject.toml
[tool.poetry]
name = "cs-ii-isp-automatic-differentiation"
version = "0.1.0"
description = "Our CS 2 ISP project!"
authors = ["Krishna Shah <kshahusa@gmail.com>", "Aryan Verma <aryanjverma2@gmail.com>", "Brennan Coil <nerblioc@outlook.com>"]
readme = "README.md"
packages = [{include = "CalCoolUs"}]

[tool.poetry.dependencies]
python = "^3.10"
networkx = "^3.1"
matplotlib = "^3.8.0"
pydot = "^1.4.2"
setuptools = "^68.2.2"
icecream = "^2.1.3"
flask = "^3.0.0"
sympy = "^1.12"
tk = "^0.1.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
coverage = {extras = ["toml"], version = "^7.3.2"}
flake8 = "^6.1.0"
flake8-bugbear = "^23.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]
[tool.coverage.report]
fail_under = 100
